refactor: rm delegate report functions
