+++ backend/main/test/groupher_server/cms/abuse_reports/changelog_report_test.exs
@@ -12,11 +12,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
 
   describe "[article changelog report/un-report]" do
     test "list article reports should work", ~m(user user2 changelog)a do
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user2)
 
       filter = %{content_type: :changelog, content_id: changelog.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = all_reports.entries |> List.first()
       assert report.article.id == changelog.id
@@ -24,10 +24,10 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
     end
 
     test "report a changelog should have a abuse report record", ~m(user changelog)a do
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
 
       filter = %{content_type: :changelog, content_id: changelog.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = List.first(all_reports.entries)
       report_cases = report.report_cases
@@ -43,11 +43,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
     end
 
     test "can undo a report", ~m(user changelog)a do
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user)
-      {:ok, _} = CMS.undo_report_article(changelog, user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.undo_article(changelog, user)
 
       filter = %{content_type: :changelog, content_id: changelog.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 0
 
       {:ok, changelog} = ORM.find(Changelog, changelog.id)
@@ -55,12 +55,12 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
     end
 
     test "can undo a existed report", ~m(user user2 changelog)a do
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user2)
-      {:ok, _} = CMS.undo_report_article(changelog, user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.undo_article(changelog, user)
 
       filter = %{content_type: :changelog, content_id: changelog.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 1
 
       {:ok, changelog} = ORM.find(Changelog, changelog.id)
@@ -70,11 +70,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
     end
 
     test "can undo a report with other user report it too", ~m(user user2 changelog)a do
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user2)
 
       filter = %{content_type: :changelog, content_id: changelog.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 1
 
       report = all_reports.entries |> List.first()
@@ -82,10 +82,10 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
       assert Enum.any?(report.report_cases, &(&1.user.login == user.login))
       assert Enum.any?(report.report_cases, &(&1.user.login == user2.login))
 
-      {:ok, _} = CMS.undo_report_article(changelog, user)
+      {:ok, _} = CMS.AbuseReports.undo_article(changelog, user)
 
       filter = %{content_type: :changelog, content_id: changelog.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 1
 
       report = all_reports.entries |> List.first()
@@ -95,11 +95,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
 
     test "different user report a comment should have same report with different report cases",
          ~m(user user2 changelog)a do
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(changelog, "reason2", "attr_info 2", user2)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason2", "attr_info 2", user2)
 
       filter = %{content_type: :changelog, content_id: changelog.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = List.first(all_reports.entries)
       report_cases = report.report_cases
@@ -113,9 +113,9 @@ defmodule GroupherServer.Test.CMS.AbuseReports.ChangelogReport do
     end
 
     test "same user can not report a comment twice", ~m(changelog user)a do
-      {:ok, _} = CMS.report_article(changelog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
 
-      assert {:error, _report} = CMS.report_article(changelog, "reason", "attr_info", user)
+      assert {:error, _report} = CMS.AbuseReports.article(changelog, "reason", "attr_info", user)
     end
   end
 end
