+++ backend/main/test/groupher_server/cms/abuse_reports/comment_report_test.exs
@@ -13,10 +13,10 @@ defmodule GroupherServer.Test.CMS.AbuseReports.CommentReport do
   describe "[article comment report/unreport]" do
     test "report a comment should have a abuse report record", ~m(user community post)a do
       {:ok, comment} = CMS.Comments.create_comment(community, :post, post.inner_id, mock_comment(), user)
-      {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
+      {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
 
       filter = %{content_type: :comment, content_id: comment.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = List.first(all_reports.entries)
       report_cases = report.report_cases
@@ -30,11 +30,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.CommentReport do
     test "different user report a comment should have same report with different report cases",
          ~m(user user2 community post)a do
       {:ok, comment} = CMS.Comments.create_comment(community, :post, post.inner_id, mock_comment(), user)
-      {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
-      {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user2)
+      {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
+      {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user2)
 
       filter = %{content_type: :comment, content_id: comment.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = List.first(all_reports.entries)
       report_cases = report.report_cases
@@ -49,8 +49,8 @@ defmodule GroupherServer.Test.CMS.AbuseReports.CommentReport do
 
     test "same user can not report a comment twice", ~m(user community post)a do
       {:ok, comment} = CMS.Comments.create_comment(community, :post, post.inner_id, mock_comment(), user)
-      {:ok, comment} = CMS.report_comment(comment, mock_comment(), "attr", user)
-      assert {:error, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
+      {:ok, comment} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
+      assert {:error, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
     end
   end
 end
