+++ backend/main/test/groupher_server/cms/abuse_reports/post_report_test.exs
@@ -12,11 +12,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
 
   describe "[article post report/unreport]" do
     test "list article reports should work", ~m(user user2 post)a do
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user2)
 
       filter = %{content_type: :post, content_id: post.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = all_reports.entries |> List.first()
       assert report.article.id == post.id
@@ -24,10 +24,10 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
     end
 
     test "report a post should have a abuse report record", ~m(user post)a do
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
 
       filter = %{content_type: :post, content_id: post.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = List.first(all_reports.entries)
       report_cases = report.report_cases
@@ -43,11 +43,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
     end
 
     test "can undo a report", ~m(user post)a do
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      {:ok, _} = CMS.undo_report_article(post, user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.undo_article(post, user)
 
       filter = %{content_type: :post, content_id: post.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 0
 
       {:ok, post} = ORM.find(Post, post.id)
@@ -55,12 +55,12 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
     end
 
     test "can undo a existed report", ~m(user user2 post)a do
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user2)
-      {:ok, _} = CMS.undo_report_article(post, user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.undo_article(post, user)
 
       filter = %{content_type: :post, content_id: post.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 1
 
       {:ok, post} = ORM.find(Post, post.id)
@@ -70,11 +70,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
     end
 
     test "can undo a report with other user report it too", ~m(user user2 post)a do
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user2)
 
       filter = %{content_type: :post, content_id: post.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 1
 
       report = all_reports.entries |> List.first()
@@ -82,10 +82,10 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
       assert Enum.any?(report.report_cases, &(&1.user.login == user.login))
       assert Enum.any?(report.report_cases, &(&1.user.login == user2.login))
 
-      {:ok, _} = CMS.undo_report_article(post, user)
+      {:ok, _} = CMS.AbuseReports.undo_article(post, user)
 
       filter = %{content_type: :post, content_id: post.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 1
 
       report = all_reports.entries |> List.first()
@@ -95,11 +95,11 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
 
     test "different user report a comment should have same report with different report cases",
          ~m(user user2 post)a do
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(post, "reason2", "attr_info 2", user2)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason2", "attr_info 2", user2)
 
       filter = %{content_type: :post, content_id: post.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       report = List.first(all_reports.entries)
       report_cases = report.report_cases
@@ -113,8 +113,8 @@ defmodule GroupherServer.Test.CMS.AbuseReports.PostReport do
     end
 
     test "same user can not report a comment twice", ~m(post user)a do
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      assert {:error, _report} = CMS.report_article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      assert {:error, _report} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
     end
   end
 end
