+++ backend/main/test/groupher_server/cms/comments/write/blog_comment_test.exs
@@ -457,11 +457,11 @@ defmodule GroupherServer.Test.CMS.Comments.BlogComment do
       {:ok, comment} =
         CMS.Comments.create_comment(community, :blog, blog.inner_id, mock_comment(), user)
 
-      {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
-      {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user2)
+      {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
+      {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user2)
 
       filter = %{content_type: :comment, content_id: comment.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
       assert all_reports.total_count == 1
 
       report = all_reports.entries |> List.first()
@@ -469,10 +469,10 @@ defmodule GroupherServer.Test.CMS.Comments.BlogComment do
       assert Enum.any?(report.report_cases, &(&1.user.login == user.login))
       assert Enum.any?(report.report_cases, &(&1.user.login == user2.login))
 
-      {:ok, _} = CMS.undo_report_comment(comment, user)
+      {:ok, _} = CMS.AbuseReports.undo_comment(comment, user)
 
       filter = %{content_type: :comment, content_id: comment.id, page: 1, size: 20}
-      {:ok, all_reports} = CMS.paged_reports(filter)
+      {:ok, all_reports} = CMS.AbuseReports.paged_reports(filter)
 
       assert all_reports.total_count == 1
 
@@ -489,7 +489,7 @@ defmodule GroupherServer.Test.CMS.Comments.BlogComment do
 
       Enum.reduce(1..(@report_threshold_for_fold - 1), [], fn _, _acc ->
         {:ok, user} = db_insert(:user)
-        {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
+        {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
       end)
 
       {:ok, comment} = ORM.find(Comment, comment.id)
@@ -505,7 +505,7 @@ defmodule GroupherServer.Test.CMS.Comments.BlogComment do
 
       Enum.reduce(1..(@report_threshold_for_fold + 1), [], fn _, _acc ->
         {:ok, user} = db_insert(:user)
-        {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
+        {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
       end)
 
       {:ok, comment} = ORM.find(Comment, comment.id)
