+++ backend/main/test/groupher_server_web/query/cms/abuse_reports/blog_report_test.exs
@@ -55,8 +55,8 @@ defmodule GroupherServer.Test.Query.AbuseReports.BlogReport do
       {:ok, blog} = CMS.Articles.create(community, :blog, blog_attrs, user)
       {:ok, blog2} = CMS.Articles.create(community, :blog, blog_attrs, user)
 
-      {:ok, _} = CMS.report_article(blog, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(blog2, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(blog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(blog2, "reason", "attr_info", user2)
 
       variables = %{filter: %{content_type: "BLOG", page: 1, size: 10}}
       results = guest_conn |> gq_query(@query, variables)
@@ -69,8 +69,8 @@ defmodule GroupherServer.Test.Query.AbuseReports.BlogReport do
       {:ok, blog} = CMS.Articles.create(community, :blog, blog_attrs, user)
       {:ok, blog2} = CMS.Articles.create(community, :blog, blog_attrs, user)
 
-      {:ok, _} = CMS.report_article(blog, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(blog2, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(blog, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(blog2, "reason", "attr_info", user2)
 
       variables = %{
         filter: %{content_type: "BLOG", content_id: blog.id, page: 1, size: 10}
@@ -91,7 +91,7 @@ defmodule GroupherServer.Test.Query.AbuseReports.BlogReport do
       {:ok, comment} =
         CMS.Comments.create_comment(community, :blog, blog.inner_id, mock_comment(), user)
 
-      {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
+      {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
 
       variables = %{filter: %{content_type: "COMMENT", page: 1, size: 10}}
       results = guest_conn |> gq_query(@query, variables)
