+++ backend/main/test/groupher_server_web/query/cms/abuse_reports/post_report_test.exs
@@ -55,8 +55,8 @@ defmodule GroupherServer.Test.Query.AbuseReports.PostReport do
       {:ok, post} = CMS.Articles.create(community, :post, post_attrs, user)
       {:ok, post2} = CMS.Articles.create(community, :post, post_attrs, user)
 
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(post2, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post2, "reason", "attr_info", user2)
 
       variables = %{filter: %{content_type: "POST", page: 1, size: 10}}
       results = guest_conn |> gq_query(@query, variables)
@@ -69,8 +69,8 @@ defmodule GroupherServer.Test.Query.AbuseReports.PostReport do
       {:ok, post} = CMS.Articles.create(community, :post, post_attrs, user)
       {:ok, post2} = CMS.Articles.create(community, :post, post_attrs, user)
 
-      {:ok, _} = CMS.report_article(post, "reason", "attr_info", user)
-      {:ok, _} = CMS.report_article(post2, "reason", "attr_info", user2)
+      {:ok, _} = CMS.AbuseReports.article(post, "reason", "attr_info", user)
+      {:ok, _} = CMS.AbuseReports.article(post2, "reason", "attr_info", user2)
 
       variables = %{filter: %{content_type: "POST", content_id: post.id, page: 1, size: 10}}
       results = guest_conn |> gq_query(@query, variables)
@@ -88,7 +88,7 @@ defmodule GroupherServer.Test.Query.AbuseReports.PostReport do
       {:ok, comment} =
         CMS.Comments.create_comment(community, :post, post.inner_id, mock_comment(), user)
 
-      {:ok, _} = CMS.report_comment(comment, mock_comment(), "attr", user)
+      {:ok, _} = CMS.AbuseReports.comment(comment, mock_comment(), "attr", user)
 
       variables = %{filter: %{content_type: "COMMENT", page: 1, size: 10}}
       results = guest_conn |> gq_query(@query, variables)
